Ajust (hoja de vida) to (to do list) componenet
refactor of hoja de vida component (to do list now)
